+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Refactor constants outside the `ReCaptcha` component.
+
 ## [1.0.9] - 2022-02-23
 
 ### Added